+++ TruePath/AbsolutePath.cs
@@ -9,7 +9,7 @@ namespace TruePath;
 /// disk letter (on Windows).
 /// </summary>
 /// <remarks>For a path that's not guaranteed to be absolute, use the <see cref="LocalPath"/> type.</remarks>
-public readonly struct AbsolutePath
+public readonly struct AbsolutePath : IEquatable<AbsolutePath>
 {
     internal readonly LocalPath Underlying;
     /// <summary>