+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] (1.0.1)
+## [1.1.0] - 2024-04-27
 ### Added
 - [#26: Publish PDB files to NuGet](https://github.com/ForNeVeR/TruePath/issues/26) (in form of `.snupkg` for now).
 - Update and publish XML documentation with the package.
@@ -43,4 +43,5 @@ This is the first published version of the package. It doesn't contain any featu
 
 [0.0.0]: https://github.com/ForNeVeR/TruePath/releases/tag/v0.0.0
 [1.0.0]: https://github.com/ForNeVeR/TruePath/compare/v0.0.0...v1.0.0
-[Unreleased]: https://github.com/ForNeVeR/TruePath/compare/v1.0.0...HEAD
+[1.1.0]: https://github.com/ForNeVeR/TruePath/compare/v1.0.0...v1.1.0
+[Unreleased]: https://github.com/ForNeVeR/TruePath/compare/v1.1.0...HEAD